convert AccountDisplay to tsx
